requirements.txt
asgiref==3.8.1
boto3==1.34.161
botocore==1.34.161
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
idna==3.7
jmespath==1.0.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
